requirements.txt
python-telegram-bot==20.7
requests==2.31.0
schedule==1.2.0